BLUEBUTTON-162 Upgrade django-oauth-toolkit to 1.1.2 main release
Update oauth references and url registration

Updates to work with django-oauth-toolkit 1.1.2

Log exception information for debugging
TODO: add request uuid

Return 404 for unauthorized read requests
Prevents knowledge of keys

Update dev env to allow unsecure callback urls

Sync python dependencies
