+++ apps/dot_ext/oauth2_validators.py
@@ -1,16 +1,10 @@
 import math
 
-from django.core.exceptions import ValidationError
-from django.utils.encoding import force_text
 from django.utils import timezone
 from django.utils.timezone import timedelta
 
 from oauth2_provider.models import AccessToken, RefreshToken
 from oauth2_provider.oauth2_validators import OAuth2Validator
-
-from oauth2_provider.validators import URIValidator
-from oauth2_provider.settings import oauth2_settings
-from oauth2_provider.validators import urlsplit
 
 
 class SingleAccessTokenValidator(OAuth2Validator):
@@ -92,27 +86,3 @@ class SingleAccessTokenValidator(OAuth2Validator):
                 access_token=access_token
             )
             refresh_token.save()
-
-
-class RedirectURIValidator(URIValidator):
-    def __init__(self, allowed_schemes):
-        self.allowed_schemes = allowed_schemes
-
-    def __call__(self, value):
-        super(RedirectURIValidator, self).__call__(value)
-        value = force_text(value)
-        if len(value.split('#')) > 1:
-            raise ValidationError('Redirect URIs must not contain fragments')
-        scheme, netloc, path, query, fragment = urlsplit(value)
-
-        if scheme.lower() not in self.allowed_schemes:
-            raise ValidationError('Invalid Redirect URI scheme: %s' % scheme.lower())
-
-
-def validate_uris(value):
-    """
-    This validator ensures that `value` contains valid blank-separated URIs"
-    """
-    v = RedirectURIValidator(oauth2_settings.ALLOWED_REDIRECT_URI_SCHEMES)
-    for uri in value.split():
-        v(uri)
