+++ requirements/requirements.dev.txt
@@ -14,13 +14,12 @@ dj-database-url==0.4.1
 django-appconf==1.0.2
 django-axes==4.1.0
 django-bootstrap-form==3.3
-django-braces==1.9.0
 django-cors-headers==1.1.0
 django-debug-toolbar==1.9.1
 django-getenv==1.3.1
 django-ipware==2.0.1
 django-localflavor==1.3
-django-oauth-toolkit-scopes-backend==0.10.0
+django-oauth-toolkit==1.1.2
 django-ses==0.7.1
 django-settings-export==1.1.0
 django-storages==1.5.1
@@ -32,7 +31,7 @@ idna==2.6
 jsonschema==2.5.1
 mccabe==0.6.1             # via flake8
 newrelic==2.106.0.87
-oauthlib==1.0.3
+oauthlib==2.1.0
 pluggy==0.6.0             # via tox
 psycopg2==2.7.4
 py==1.5.3                 # via tox
